API data 리포지토리 테스트 성공!

+++ src/main/java/com/example/demo/movies/repository/MoviesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.movies.entity.Movies;
 
-public interface MoviesRepository extends JpaRepository<Movies, Integer>   {
+public interface MoviesRepository extends JpaRepository<Movies, Long>   {
 
 }
